# frontend/openapi/admin-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "REST to MCP Admin API",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/backend-resources/": {
      "get": {
        "tags": [
          "backend-resources"
        ],
        "parameters": [
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "in": "query",
            "name": "page",
            "required": false
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "in": "query",
            "name": "pageSize",
            "required": false
          },
          {
            "schema": {
              "type": "string"
            },
            "in": "query",
            "name": "search",
            "required": false
          },
          {
            "schema": {
              "type": "boolean"
            },
            "in": "query",
            "name": "isActive",
            "required": false
          },
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "query",
            "name": "backendApiId",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "backendApiId": {
                            "type": "string",
                            "format": "uuid"
                          },
                          "name": {
                            "type": "string",
                            "minLength": 1
                          },
                          "operationId": {
                            "type": "string",
                            "minLength": 1
                          },
                          "description": {
                            "type": "string",
                            "nullable": true
                          },
                          "httpMethod": {
                            "type": "string",
                            "enum": [
                              "GET",
                              "POST",
                              "PUT",
                              "PATCH",
                              "DELETE"
                            ]
                          },
                          "pathTemplate": {
                            "type": "string",
                            "minLength": 1
                          },
                          "bodyTemplate": {
                            "type": "string",
                            "nullable": true
                          },
                          "requestSchema": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "responseSchema": {
                            "type": "object",
                            "additionalProperties": {},
                            "nullable": true
                          },
                          "isActive": {
                            "type": "boolean"
                          },
                          "createdAt": {
                            "type": "string",
                            "format": "date-time"
                          },
                          "updatedAt": {
                            "type": "string",
                            "format": "date-time"
                          }
                        },
                        "required": [
                          "id",
                          "backendApiId",
                          "name",
                          "operationId",
                          "httpMethod",
                          "pathTemplate",
                          "requestSchema",
                          "responseSchema",
                          "isActive"
                        ],
                        "additionalProperties": false
                      }
                    },
                    "meta": {
                      "type": "object",
                      "properties": {
                        "pagination": {
                          "type": "object",
                          "properties": {
                            "page": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "pageSize": {
                              "type": "integer",
                              "exclusiveMinimum": true,
                              "minimum": 0
                            },
                            "total": {
                              "type": "integer",
                              "minimum": 0
                            },
                            "pageCount": {
                              "type": "integer",
                              "minimum": 0
                            }
                          },
                          "required": [
                            "page",
                            "pageSize",
                            "total",
                            "pageCount"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "required": [
                        "pagination"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data",
                    "meta"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "backend-resources"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "backendApiId": {
                    "type": "string",
                    "format": "uuid"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "operationId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "httpMethod": {
                    "type": "string",
                    "enum": [
                      "GET",
                      "POST",
                      "PUT",
                      "PATCH",
                      "DELETE"
                    ]
                  },
                  "pathTemplate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "bodyTemplate": {
                    "type": "string"
                  },
                  "requestSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "responseSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "required": [
                  "backendApiId",
                  "name",
                  "operationId",
                  "httpMethod",
                  "pathTemplate"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "backendApiId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "operationId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "httpMethod": {
                          "type": "string",
                          "enum": [
                            "GET",
                            "POST",
                            "PUT",
                            "PATCH",
                            "DELETE"
                          ]
                        },
                        "pathTemplate": {
                          "type": "string",
                          "minLength": 1
                        },
                        "bodyTemplate": {
                          "type": "string",
                          "nullable": true
                        },
                        "requestSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "responseSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "backendApiId",
                        "name",
                        "operationId",
                        "httpMethod",
                        "pathTemplate",
                        "requestSchema",
                        "responseSchema",
                        "isActive"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      }
    },
    "/backend-resources/{id}": {
      "patch": {
        "tags": [
          "backend-resources"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "operationId": {
                    "type": "string",
                    "minLength": 1
                  },
                  "description": {
                    "type": "string"
                  },
                  "httpMethod": {
                    "type": "string",
                    "enum": [
                      "GET",
                      "POST",
                      "PUT",
                      "PATCH",
                      "DELETE"
                    ]
                  },
                  "pathTemplate": {
                    "type": "string",
                    "minLength": 1
                  },
                  "bodyTemplate": {
                    "type": "string"
                  },
                  "requestSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "responseSchema": {
                    "type": "object",
                    "additionalProperties": {},
                    "default": {}
                  },
                  "isActive": {
                    "type": "boolean",
                    "default": true
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "backendApiId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "name": {
                          "type": "string",
                          "minLength": 1
                        },
                        "operationId": {
                          "type": "string",
                          "minLength": 1
                        },
                        "description": {
                          "type": "string",
                          "nullable": true
                        },
                        "httpMethod": {
                          "type": "string",
                          "enum": [
                            "GET",
                            "POST",
                            "PUT",
                            "PATCH",
                            "DELETE"
                          ]
                        },
                        "pathTemplate": {
                          "type": "string",
                          "minLength": 1
                        },
                        "bodyTemplate": {
                          "type": "string",
                          "nullable": true
                        },
                        "requestSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "responseSchema": {
                          "type": "object",
                          "additionalProperties": {},
                          "nullable": true
                        },
                        "isActive": {
                          "type": "boolean"
                        },
                        "createdAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "updatedAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "id",
                        "backendApiId",
                        "name",
                        "operationId",
                        "httpMethod",
                        "pathTemplate",
                        "requestSchema",
                        "responseSchema",
                        "isActive"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "backend-resources"
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "in": "path",
            "name": "id",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "/api/admin/v1"
    }
  ]
}